refactor(benchmarks): decouple the CI baseline from the local analytics layer
Prep for splitting the local benchmarking (memory, sweep, plotting, …) into a
standalone package — makes the CodSpeed CI baseline import-clean of the
analytics layer:

- Move `spec_param_id` (the `<name>-<axis>=<value>` param-id convention) from
  `snapshot.py` into `registry.py`. It's used by the core (`registry.param_ids`,
  `test_to_solver`) and was the only core → analytics import.
- Stop eagerly importing `bench` in `benchmarks/__init__`; `models`/`patterns`
  stay (they register specs). `from benchmarks import bench` still works on
  demand (submodule import).

Result: importing the package + the three CodSpeed test modules pulls zero of
{memory, sweep, plotting, snapshot, bench, cli}; the dependency arrow now only
points analytics → core. No behaviour change — collection, harness tests (incl.
the memory↔pytest id alignment), and the memory CLI are all unaffected.
