+++ benchmarks/test_to_solver.py
@@ -20,8 +20,7 @@ import pytest
 
 from benchmarks.conftest import maybe_skip
 from benchmarks.phases import SOLVER_HANDOFFS
-from benchmarks.registry import ModelSpec, iter_params
-from benchmarks.snapshot import spec_param_id
+from benchmarks.registry import ModelSpec, iter_params, spec_param_id
 from linopy.solvers import available_solvers
